Add `Get-SecureRandom` cmdlet
